refactor(map): own growth and bundle adjustment
